Move Selection code into Apollo.Selection
