+++ Apollo/Components/FrameDisplay.cs
@@ -15,7 +15,7 @@ using Avalonia.VisualTree;
 using Apollo.Core;
 using Apollo.Devices;
 using Apollo.Elements;
-using Apollo.Interfaces;
+using Apollo.Selection;
 using Apollo.Structures;
 
 namespace Apollo.Components {
