+++ Apollo/DeviceViewers/GroupViewer.cs
@@ -17,7 +17,7 @@ using Apollo.Core;
 using Apollo.Devices;
 using Apollo.Elements;
 using Apollo.Helpers;
-using Apollo.Interfaces;
+using Apollo.Selection;
 using Apollo.Viewers;
 using Apollo.Windows;
