+++ Apollo/Devices/Choke.cs
@@ -2,7 +2,7 @@ using System.Collections.Concurrent;
 
 using Apollo.DeviceViewers;
 using Apollo.Elements;
-using Apollo.Interfaces;
+using Apollo.Selection;
 using Apollo.Structures;
 
 namespace Apollo.Devices {
