+++ Apollo/Devices/Pattern.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using Apollo.Elements;
 using Apollo.Enums;
 using Apollo.Helpers;
-using Apollo.Interfaces;
+using Apollo.Selection;
 using Apollo.Structures;
 using Apollo.Windows;
