+++ Apollo/Elements/Project.cs
@@ -12,7 +12,7 @@ using Apollo.Binary;
 using Apollo.Core;
 using Apollo.Enums;
 using Apollo.Helpers;
-using Apollo.Interfaces;
+using Apollo.Selection;
 using Apollo.Windows;
 
 namespace Apollo.Elements {
